massive commit going to agent.link.Message.Args as a []ByteSlice

+++ src/coord/state_transform.go
@@ -21,7 +21,7 @@ type StateTransform struct {
 
 func transformFromState(state *cagent.AgentState) *StateTransform {
     self := new(StateTransform)
-    self.turn = state.Turn
+    self.turn = uint64(state.Turn)
     self.pos = state.Position
     self.energy = state.Energy
     self.alive = state.Alive
